MDEV-13275:Query optimizer not picking optimal ORDER BY PRIMARY in case of INNER JOIN on PRIMARY like it does for similar WHERE condition
Currently the code for propagating equalities in the ORDER BY columns
does a full substitution (that is x=y, then f(x) = f(y)).
For the ORDER BY clause, we can allow equality propagation even when full substitution is
not possible because with the ORDER BY clause we do ONLY comparisons. Also the equality
propagation for columns in the ORDER BY clause are done only for top level Item_fields.